hpcprof: Remove dead code
numMetricsPerGroup() is to compute the number of metrics per group. It is used by hpcprof to find a metric index after metric "transformation" (like after creating summary metrics).
If we want to compute derived metrics in hpcprof, we'll need to compute the metric index after each "transformation".
However, this function uses heuristic approach and can be fragile to any modification in hpcprof. It's better to compute derived metrics in hpcviewer.

Minor update: README