hpcrun: use '#' sign instead of '$' in metric formula
hpcviewer now recognizes three formula modifiers:
- '@' prefix to point to the aggregate of a metric
- '$' prefix to point to the scope value of a metric
- '#' prefix to point to the scope value of a metric based on hpcrun's "metric order"

hpcrun needs to avoid using '$' in the formula because it has no knowledge of what a metric index will be.
However, it knows the "order" of the metric which is not the same as the "index" of the metric.
The later is rearranged by hpcprof after manipulating with temporary metrics and who knows what the index will be.
The former is more consistent and hpcprof keeps track the order of the metric and passes it to hpcviewer.

This commit replaces '$' prefix with '#' in memleak and gpu-metrics