Refine metric visibility
Usually we only have 2 values of visibility: true (show) and false (hidden).
Now we refine the visibility field into 4 values:

0: hide
1: show
2: show the inclusive one
3: show the exclusive one

For the sake of backward compatibility, 0 and 1 have to be "hide" and "show".

This "small" modification has big impact in hpcprof since in many parts, we used to handle the boolean type for visibility (either show or hide).