hpcprof: Add a procedure type to be invisible in hpcviewer
Some place holders or procedures should not be visible in hpcviewer.
hpcprof has to specify which procedures have to be hidden in hpcviewer, but can be shown in hpctraceviewer.

This commit uses the existing attribute "f" in Procedure tag. Example:

    <Procedure i="450444" n="&lt;no activity&gt;" f="3" v="0"/>

In this case, procedure "<no activity>" (the value of attribute f is 3) will be hidden in hpcviewer.
See constants in lib/prof/NameMappings.cpp for possible values of attribute f.