refinements to post-processing to improve hpcviewer's view (#281)
* hpcprof: Add a procedure type to be invisible in hpcviewer

Some place holders or procedures should not be visible in hpcviewer.
hpcprof has to specify which procedures have to be hidden in hpcviewer, but can be shown in hpctraceviewer.

This commit uses the existing attribute "f" in Procedure tag. Example:

    <Procedure i="450444" n="&lt;no activity&gt;" f="3" v="0"/>

In this case, procedure "<no activity>" (the value of attribute f is 3) will be hidden in hpcviewer.
See constants in lib/prof/NameMappings.cpp for possible values of attribute f.

* Add description in DTD

Add missing description and possible values in XML attributes

* hpcprof: set GPU's *_ACUMU metrics to be invisible

Accumulator metrics should be invisible to users since they are used for internal usage as helper metrics.

* hpcprof: fix gpu_trace place holder

Instead of renaming gpu_trace into <gpu kernel>, we  should rename it as <gpu trace>. Plus, we need to make it invisible in hpcviewer.

* hpcprof: rename gpu_op_trace to <gpu kernel> back

Both gpu_op_kernel and gpu_op_trace are renamed back to <gpu kernel> with differences:

- gpu_op_kernel is visible both in hpcviewer and hpctraceviewer (in theory)
- gpu_op_trace  is only visible in hpctraceviewer but NOT in hpcviewer because it isn't associated with any metrics. It is used in GPU sample-source for traces only.

To accommodate this, NameMapping.cpp has to be refined so that in experiment.xml :

- gpu_op_kernel has attribute f="1"
- gpu_op_trace  has attribute f="3"

* hpcprof: add a new type for place holders

Add a new type of place holder: Top-down place holder. This place holder is only visible in top-down view.
Example: <program root>, <thread root>, ...

Co-authored-by: Laksono Adhianto <laksono@gmail.com>