Rework the gprof options to make them conditional on a command-line option.
By default, do nothing.  The option --disable-gprof for both hpcrun
and hpclink adds overrides to four gprof functions and turns them into
no-ops: __monstartup, _mcleanup, mcount and _mcount.

Hpcrun and gprof (-pg) won't run together and result in a segfault.
It would be better to recompile the binary without gprof.  But if
that's not possible, then this option provides another way to disable
gprof and allow hpcrun to run.