Step one of hiding hpcrun's functions and libraries.  This much works (#317)
* Step one of hiding hpcrun's functions and libraries.  This much works
around the deadlock with openmpi 4.x and UCX on stria at SNL.

  1. configure.ac tests if libunwind and lzma have a static (.a)
  archive compiled with -fPIC.  If yes, then proceed, else issue a
  warning and build libhpcrun.so as we previously do.

  2. Copy libunwind.a and lzma.a into src/extern and use objcopy to
  --redefine-sym mmap() and munmap() into hpcrun_real_mmap and
  hpcrun_real_munmap.  Using a separate directory is an easy way to
  make sure that these are built before hpcrun.

  3. In hpcrun/extern-real, write hpcrun_real_* functions to call the
  real versions via syscall().

  4. Edit hpcrun/Makefile.am, hpcrun, hpclink to allow both old and
  new versions of linking.

Remove the --enable-libunwind-primary option and clean up the related
hash-defines.  Libunwind is now too intertwined in the rest of the
code to be optional.

Not yet done: localize the symbols in hpcrun, and stripping the line
map without stripping the EH frames.

* Adjust the setting of LIBUNWIND_COPY to make it more clear when we
need to copy libunwind (.so exists and .a does not have fpic).