Allow disabling the LD_AUDIT-based infrastructure using --disable-rtld-audit
Using this flag allows for hpcrun to still function on systems where LD_AUDIT
support in Glibc is currently non-functional. There will be R*PATH issues when
this is used, similar to the state of hpcrun previous to all this.

Note that with both this and the previous commit, libmonitor should be configured
with --disable-dlfns, regardless of the --enable-rtld-audit setting.