Delay auditor notifications until libhpcrun is certainly ready
If libhpcrun initializes early, the auditor may witness its initialization and assume its prepared to recieve notifications before it actually is (via a recursive initialization attempt).
This patch instead assumes that the mainlib will notify the auditor when it is prepared, and thus delays notifications until that point.

The middling `connecting` state is still needed to prevent (some) recursive initializations.

Fixes issues with PeleC and other applications that spawn threads in their init constructors.