Fix issue 356 -- CPU time is overcounted when threads sleep. The code used the delta in REALTIME as the value of CPUTIME when profiling by the latter.  If a thread was sampled at T-epsilon, and then went to sleep until T2 + epsilon, the time between T2 and T would be incorrectly said the be CPU time.
The fix is to read the CPU time for the delta when profiling CPUTIME,
and the real time for the delta when profiling REALTIME.

While I was in there, I deleted the various tortuous bits of code that were
needed to support BlueGene, which is now completely obsolete.  I also deleted
the profiling of WALLCLOCK, which was just an alias for CPUTIME, and profiling
by the undocumented timer ITIMER.

Reworked the hpcrun -l message to correctly describe the timer options.