fix opencl wrappers so they don't always try to measure
opencl wrappers assumed that -e gpu=opencl was provided
as a command line argument. if that wasn't done, the
opencl wrappers would call through uninitialized function
pointers.

if any opencl wrapper is used, make sure that all of the
function pointers to the wrapped functions are bound
before any is used.