fix issue 321: dynamic library unmap notification fails
issue 321 reports that synprog so.sx.so fails.

the root cause: hpcrun_loadmap_notify_unmap
was called too late - after dso_info for a load module has
been cleared. the notification must be performed before
dso_info is cleared because it needs to clear the unwind recipe
map for the address range of the dso, which is only known before
dso_info is cleared.

1. fix issue 321 by adding notification before clearing dso_info

2. commit hpcrun_loadmap_dump (and helper functions) written to
   help debug this issue.