Redesign the MPI ordering constraints to be more strict and better all around.
Now the Sinks and Sources report and enforce their own ordering requirements, rather than the Pipeline builder having to specify dependencies.
The exact order is based on insertion order, so similar codes will generate identical orderings.

As of now, all MPI usage should be through the `mpi::` sanitization wrappers and be either outside of a Pipeline or guarded by an ::enterOrdered* call.