Initial attempts at the new "collaborative" Context possibilities
With some measurements, we may know "where" something happened but not precisely "who" did it.
The CollaborativeContext represents this as a series of "collaborating" subtrees embedded in
the normal Context tree, for which metrics are summed and redistributed back to the "collaborating"
Threads. Developed for the new CUDA measurement support.

Note that for obvious reasons this increases the memory theoretical bound significantly.
This cannot be avoided with this algorithm so we just try not to raise the bound _every_ time.

Note also that this currently orders the Superposition distribution prior to the Collective
distribution. Since none of this actually works yet its unknown whether that's a problem.

Also, don't try with multiple ranks, nothing will work there I guarentee.