Tweaks and reductions for the dependency management framework
This separates out the headers' dependencies from the link-time dependencies listed in the `static_library` itself,
taking advantage of the dependency chains to make simpler to use downstream and reducing the leaf dependency lists.
This also removes a few dependency links, it seems to compile fine in a clean-room but further testing is needed.

Its probably overkill, but it makes sense for the current organizational structure.