fix placeholder renaming
if libhpcrun and libmonitor are stripped, hpcstruct transforms
function names to include the load module. this gets in the
way of placeholder renaming. change renaming to consider
two names equivalent if they are the same, at least up to the
first space if one name is longer than the other.