Fix GDB's confusion over our use of clone()
GDB "loses" its breakpoints during the Fnbounds server launch for *very* unclear reasons,
it appears to be a bug reproducable in 20 LoC. This should probably be reported upstream,
for the immediates though this solves the issue.

First, the GDB bug only triggers for "fork-like" clones (where `SIGCHLD` is set), so the
return signal is removed from both clones. This requires a change to the waitpid calls to
support acting on something that didn't come out of a fork() (no function change).
This also has the side effect of removing the "loose" `SIGCHLD` signal(s) that get caught by
libmonitor. The less crazy recursion the better.

Second, without `SIGCHLD` GDB considers the child and grandchild "threads" of the parent, so when
the grandchild exec's it presumes the entire process has been nuked (and further chaos ensues).
Adding `CLONE_UNTRACED` to the grandchild clone makes GDB (forcably detach and) ignore the
grandchild process completely.

Although it seems harmless to leave it out on GDB 10.1, this commit also adds `CLONE_UNTRACED`
to the child clone to remove any chance for more issues. I just don't have enough confidence in
GDB's buggy clone handling.