Avoid hanging at exit when using CLONE_VM on Glibc <2.24
There is a bug in older Glibc (https://sourceware.org/bugzilla/show_bug.cgi?id=18862) where using `CLONE_VM`
instead of `CLONE_THREAD` causes the Glibc wrapper to reset its internal pthread state in the parent process.
This wreaks havoc with all things pthread, but most noticably it stops the `pthread_kill` in libmonitor's
shootdown from functioning (and libmonitor doesn't pay attention enough to the returned error codes).

The fix here is to remove `CLONE_VM`, which as a side effect removes the parent's tracking for the Fnbounds
process. Although there is a "shutdown" message that could be sent to close Fnbounds, since sending the
message may block indefinitely this is not used to signal a shutdown (unchanged from the original). In the
worst case scenario of an Fnbounds hung between queries, the application will exit without concern even
though Fnbounds has not yet (previously it would `waitpid` for it to exit). Caveat emptor.

Since this bug only affects the parent, the child and grandchild share a memory space. Neither use pthreads
and both exit quickly (and through different paths) so it should be fine.

Note that this will re-raise the slowdowns observed with large memory spaces. Caveat emptor.