Remove recursion during CCT walks
The algorithm used to walk the CCT use simple recursive algorithms, for complex applications
this complicated debugging with hundreds of identical frames.

This patch replaces it with a non-recursive algorithm using a variable in the nodes themselves to
store state. This should be as thread-safe as the previous algorithm, just without the recursion.

hpcrun_cct_merge and related still use a recursive algorithm, they are left effectively
untouched in this commit.