QoL: Walk CCTs without using recursion
The algorithm used to walk CCTs in hpcrun is implemented using simple
recursive algorithms. While these algorithms work (most of the time),
they also run a high risk of running out of stack space at the most
critical time of the process: when writing out the data. The stack in
these algorithms can be many times deeper than the application's stack
ever was, and debugging errors with such a deep callstack is a painful
enterprise.

This commit replaces the simpler algorithms (i.e. those that do not
"merge" CCTs) with a non-recursive variant of the same traversal.
In place of the process stack, a new pointer field is added in each cct
node which allows them to form a singly-linked stack. This field is
updated as the traversal progresses, while the traversal itself is
implemented as simply nested loops.

Note that, unlike the recursive algorithm, this means only one traversal
can occur concurrently within a CCT. This means any CCT shared between
threads (e.g. for GPU support) must not walk the CCT concurrently, in
addition to not writing when inappropriate. Caveat emptor.

The `hpcrun_cct_merge` functions are left untouched in this commit.