Mark the entire CUPTI subscriber callback as "safe"
When monitoring both CPU and GPU metrics, its possible that Run will non-deterministically
hang trying (and failing) to gather CPU metrics for the CUPTI subscriber callback.
The log file in this case contains a number of failed unwinds coming from this function.

This patch wraps the entire callback in a hpcrun_safe_enter/exit pair, preventing
samples in the region and (it seems) solving the hang.