Rewrite the ExperimentXML4 Context output
This is a first pass to fix the vast majority of issues with trace rendering.
I don't think this is a complete fix, there are quirks that I don't fully understand
the need for, and the longevity of this approach is questionable due to the difference
in design between EXMLv4 CCTs and Prof2 Contexts. The remaining concerns and
details will be the subject of a new GitHub issue.

Also adds some much-needed documentation for the Scope semantics, and adjusts
the DTD to fit the additions from the new EXMLv4 format.