Implement first-class placeholder support. Resolves https://github.com/HPCToolkit/hpctoolkit/issues/461
There are many common scenarios where the libhpcrun.so or libmonitor.so used to
perform measurement disappears or becomes out-of-date before Prof is run.
This causes all "placeholder" contexts to show up as (at best) <unknown procedure>
or (at worst) other random and unrelated functions. The affected placeholders include:
 - GPU operation placeholders (<gpu *>)
 - OpenMP operation placeholders (<omp *>)
 - <no activity>
 - CPU fences (<program root> and <thread root>)

This patch refactors and unifies all placeholders across Run into prof-lean/placeholders.{h,c},
which are now recorded as offsets in the special load module 0. This load module is never
associated with a real binary, thus making placeholders unaffected by changes
in the Toolkit installation.

For an additional touch of compatibility, the offsets used for placeholders are
not contiguous and are instead comprised of 8-byte "shortcodes" (strings). These
(vaugely readable) codes are used to generate fallback names for unrecognized
values, allowing measurements to be passed between different versions of the
placeholder set without losing easy interpretation by developers.