Implement K-disordered timepoint (trace) sorting in the Pipeline
There are cases where the trace output from Run will not be in perfect order due to measurement constraints.
We expect that these cases are only "slightly" disordered, so we don't have to load the whole trace to sort
but instead can use a "streaming-sort" based on a maintained heap. This algorithm is added as a new
abstraction for use elsewhere (e.g. the heap for the CMS output algorithm).

The Pipeline is in charge of sorting timepoints as they flow past, the interface now will indicate whenever
the Source and Sink(s) need to rewind due to a sorting issue (i.e. falling back to a larger sort window).
The Source is also in charge of reporting the "expected disorder," which is used to govern the initial
sorting window size. If the "disorder" is 0 the Pipeline will expect fully ordered timepoints.

For simplicity, this commit also removes the overloads for timepoints with missing arguments; now all
timepoints have a Thread, Context and timestamp, period. The "timepoint bounds" case (for min/max times)
is now handled explicitly as a part of the attributes DataClass, but it only stablizes after write().
(Which probably doesn't make a lot of sense, but it works so its fine.)