Optimize the output of timepoints data
Previously Sinks would be notified for every single timepoint as it was emitted, at
the cost of virtual function call overheads, reduced optimization potential, hot
fine-grained synchronization, etc. This rewrites the interface to instead pass
timepoints in groups (currently <4096), which allows for better algorithms in the
HPCTraceDB2 itself.

The main new optimization is the addition of a very small (only 2 entries)
on-the-stack cache for reducing queries to the ragged_vectors for Contexts.
Why only 2? Applications move fast. For a miniature LAMMPS dataset:
 - Top cache entry (MRU) gets hit 75-99% of the time,
 - Of what remains, second cache entry gets hit ~2% of the time (at best),
 - Of what remains, <0.7% gets hit *at all* in a 64-sized cache.

Bigger caches have diminishing returns for the hardware cache workout, and the cost
of a miss is mostly just a hardware case miss and coherence, so 2 entries is plenty.

Of course, the linear static-allocation LRU cache is a new abstraction. In case it
needs to be used anywhere else.