Check trace orderings and decide disorder in Run
Instead of the boolean "is ordered" NV flag, Run now passes an integer for "expected maximum disorder."
This value matches the semantic for disorder for the Prof2 interfaces, Hpcrun4 now passes it through.

In Run, all traces are now checked for correct order. If a trace is ordered it is marked with a disorder
of 0 (indicating no need to sort), otherwise the disorder is a value decided during the initialization
of the trace. Currently this is 5 for most traces and 30 for GPU traces.

Note that since Prof2 has fallbacks when the expected disorder was surpassed, there won't be a
correctness issue if these are values turn out to be too small (although there may be a performance cost).