Fix Prof2 hostid handling. Fixes https://github.com/HPCToolkit/hpctoolkit/issues/469
In some (rare) cases gethostid() doesn't generate a consistent value. Most of the calls
to gethostid() go through OSUtil_hostid(), which has an internal process-wide cache.
This replaces the direct calls to gethostid() added for Prof2. Without this, individual
threads/streams may be relocated to a different NODE than the rest of their MPI rank.

In addition, on 64-bit machines gethostid() returns a signed 64-bit value, of which only
the lowest 32 bits are useful. These two lines contain an implicit cast to uint64_t which
retains the sign-extension, creating very large (negative) NODE values on systems with
hostids starting with >=0x8. OSUtil_gethostid() handles this oddity internally, so this
patch also fixes this issue.