Fixup OSUtil_hostid for thread-safety and unsigned return
Doesn't fix any known issues, but may prevent very hard-to-debug issues in the future.

There was a comment in this region that claimed some versions of RHEL 6.x had issues
with gethostid() triggering a SIGFPE, the previous implementation would catch and
gracefully absorb the signal. There is no clean and contained way to do this in a
thread-safe manner, so this version doesn't do so (and thus is much simpler).