Refactor "timepoints" to "ctxTimepoints," specifying Context-type timepoints
The idea here is that "timepoints" is a general term for the set of [Thread,
timestamp, *] tuples. For Context-type timepoints, the additional element
is a Context, making them execution traces. There will be other "timepoints"
in the future, so this refactor makes it more obvious which type of timepoint
we're dealing with at (pretty much) every stage of the process.

It seems unlikely that any type of timepoint will ever be less than or
exactly [Thread, timestamp], Threads represent (vaugely) "physical" execution
locations, tracking only the times at which such a location "exists" is of
little use for performance analysis (e.g. adding the Context which the Thread
was executing at the given time is far more useful). The new snippets of template
magic are written under this assumption.

No actual functional change, but needed for the following commit.