Add metric-type timepoints
Metric-type timepoints include a Metric and its measured value, in addition
to the usual Thread and timestamp. This makes them suitable for traces for
system-wide measured values that may change over time.

The new interface is not currently used and thus is completely untested, but
this is a needed addition to the core for https://github.com/HPCToolkit/hpctoolkit/issues/442.
The template magic is used for both types of timepoints, hopefully this
means the result will be relatively bug-free.