Fix Structfile parsing of filenames
Apparently the `f=` attribute is optional on most tags, and Struct takes advantage
of this to elide some filenames in double-alien arrangements. This causes Files
with empty names to be emitted into the Pipeline, which then of course causes
confusing results in the final database.

This commit adds some extra checks to ensure empty filenames instead use the
parent's file (as they should) and adds an assertion in the File constructor to
ensure something like this doesn't happen again.