Move sub-Metric ids into the Prof2 core
ExperimentXML4 and SparseDB both had code to agree on the ids used for
StatisticPartial values, which looked like weird `<< 8 + idx` expressions.
While it works, its not exactly clean or convenient to maintain.

This commit removes those expressions and instead implements the pieces
in the core to support ids for StatisticPartials. See Metric::Identifier
for details.