Rewrite the entire Context class framework
Before this commit, calling context reconstruction was based on the idea of attributing to
generic "Context-likes," where most of the code would be agnostic to whether a Context-like
was an actual full Context or was a SuperpositionedContext. With range-based PC sampling
this got more complex (ContextReconstructions don't have parents), certain Context-like types
didn't make sense in various places, and massive switch statements grew in various places in
the code. Buggy and very hard to describe in useful terms.

This commit rewrites the entire framework to be type-safe (no std::variants) and reducing
it down to exactly 3 representations of calling context:
 - Contexts, for the usual full calling Context (A -> F - X -> B),
 - ContextReconstructions, for Contexts missing their nearest parents (A -> F - ... -> B), and
 - ContextFlowGraphs, for Contexts missing all their parents (... -> B).
Metric values attributed to the latter two are redistributed to Contexts by generating
"factors" for the various options, same as before. Unlike before, there's more comments.

Before this commit, range-based PC sampling required the idea of "delaying" Thread-local
metric values until the Collaboration redistribution happened. This increased the memory
bound by quite a bit since the CPU metric data also needed to be kept around until then.
With this commit there is no cross-Thread redistribution, instead Threads can be created as
"merged Threads" which are shared between multiple ProfileSources (normal Threads are
Source-local) and uniqued by the id-tuple. GPU data is attributed to these and then
redistributed inside, CPU data is processed like normal. Lower memory footprint and much
easier to maintain overall.

This commit contains a bunch of churn to keep the normal Contexts working for the most part,
but the new Contexts do not have support in the MPI bits. The MPI side requires more effort
to sort out so that's left for a later PR. So long as you don't use PC sampling with
hpcprof2-mpi everything will work as expected.