Add a new flag to CCT nodes to indicate when the parent-child link is in need of reconstruction
GPU samples can't be unwound, so indicate that in the CCT instead of relying on information in the
Structfile or binary path. By generalizing this we can give an explicitly poor answer when
CFGs are available (or are too complex for us). And it helps keep everyone aware of what's happening
in Prof2, which is the big reason to do this.

This might be useful for interpreting flat CPU samples too. Not sure though. Kinda a different beast.

This is an incompatible change to the .hpcrun 4.0 format, so all measurements need to be redone.