Remove ProfileTransformer and replace it with a refined ProfileFinalizer
Transformers were part of the original conception of Prof2 as a "pipeline-based"
framework: data comes out of the Sources, gets changed by the Transformers, extra
data is attached by the Finalizers, and the whole gets consumed by the Sinks.
At a high level this gives a nice flow of data through the program.

Over time the need for Transformers lowered quite a bit, there was only 4
Transfomers and one was abusing the interface for MPI reasons. All of the data
is in the Classification which is generated by the Finalizer(s).

Also, both Transformers and Finalizers run on the idea that "everyone gets a
chance to edit the answer." In practice that requires the top-level caller (hpcprof2)
to know how each Transformer or Finalizer will edit the answers and order them
accordingly. That makes the top-level code more sensitive than it really should be.

This commit does three things: first, it removes Transformers completely, shifting
all the responsibilities onto the Finalizers. ExtensionClass::classification now
also includes the application of Classification data, and the new
ExtensionClass::statistics contains the addition of Statistics to a Metric.

Second, it removes the common Classification userdata storage, instead requiring
Finalizers to register a custom userdata for it for themselves. This keeps the
Finalizers from overwriting other classification data from other Finalizers.

Third, it refines the Finalizer interface to remove most cross-Finalizer interactions.
For every ExtensionClass Finalizers now give their "final answer" if they have one,
or std::nullopt if they don't. The answer from the first Finalizer that gives one is
used, if no Finalizer gives an answer a reasonable default is used instead. In particular,
Context creation is no longer deeply recursive (!).

(The exception to the above is Classifications, which are mostly untouched to be handled
in a subsequent commit.)