Factor Function and File out into a new source file
No functional change to Files other than the new header.

Functions are now significantly more intellegent, with an operator+= overload for
merging Functions in a commutative manner. They still aren't allocated by the
Pipeline or have userdata, the intention is for Finalizers to allocate them as part
of their own Module userdata members. To support this they allow all kinds of motion.