Add util::interval template type
Intervals represent half-open intervals on an arbitrary linear space. The default
comparisons define a partial order where overlapping intervals are incomparible
(a < b && b < a && a != b). Using them as keys in a std::map creates an interval
lookup table (any lookup within an interval returns the value for the whole interval).