Initial implementation of meta.db support (Sink and hpcproftt)
The low-level conversion functions for meta.db use a new approach, instead of
using FILE* and fread/fwrite style interfaces, the new functions convert
between structures and byte-arrays instead. This abstracts away the I/O from
the conversion, making it easier to use the appropriate I/O for the situation.

These functions and structures may be revised later for performance, if the
memory copies become a bottleneck.