Fix a SEGV in StructFile parsing
Xerces-C++ requires that all usage be surrounded by a pair of calls to
XMLPlatformUtils::* to manage some global internals, but because of the
destructor order there was a case where sometimes the destructor would
be called outside of that pair.

This commit instead binds XMLPlatformUtils to the static constructors.
Not great, but it works.