Refactors to fully fix inlined Contexts and prepare for meta.db
There is technically an issue with the metric values for a line containing
an inlined function call: since inlined_function also contains the call line,
it is not a child of the line Context and thus the line Context has a lower
MetricScope::execution ("inclusive") metric value than it should. It doesn't show
up since the Viewer only presents MetricScope::function ("exclusive") values
for line Contexts, but it's been there for a (long) while.

This commit fixes this issue (and hopefully those like it) by revisioning how
calls are represented: instead of relying on a (complex) function of the
parent and child Scopes, the relation is explicitly listed as a Relation
enum value. This gives Finalizers full flexibility to write Scopes however
they view the world, while still providing the key properties that are needed
for, say, metric propagation to come out right.

With the new structure, inlined_function Scopes are now removed and written
as line Scope + Relation::inlined_call + function Scope. This splitting is
a necessary change for the current meta.db proposal, since it requires those
two Scopes to have separate ids.

Notes:
 - The old Scope-based structure still works, so EXML hasn't been updated to
   read Relations properly. It likely won't be before it's purged in favor
   of meta.db.
 - An attempt is made in this commit to update MPI bits to work with Relations,
   it's basically untested since MPI is still borked in places. It's at the
   bottom of my high-priority list.