Implement formulas as a proper Expression tree
Formulas have been represented as C-like expressions for a long while,
but 1) the details are documented only in source files in the Viewer
repo, 2) any substantial inspection/adjustment of formulas is more painful
than it should be, and 3) doing this limits our output formats to those
that use C-like expressions.

This commit replaces formulas with proper Expression trees, which is
required for the upcoming METRICS.yaml format (where math is written
in the YAML structure instead of straight strings).