Improve the ParallelFor helper templates
ParallelForEach now allows the action to be set during the call to fill. Setting this later
allows for data to be saved in the callable itself, saving a few steps and making
lambda-based loops more useful.

The new template ParallelFor is the same as ForEach, except instead of items in a
std::vector the items are just indicies starting from 0. This simplifies the cases where
the data is already stored somewhere and only the work needs to be distributed.

Still rough around the edges, but it works for now.