Replace asserts with aborts
assert supports NDEBUG, so it will compile out if NDEBUG is defined. Our use
of Autotools doesn't define it, but Meson does. This commit swaps the asserts
that seem to be based on external factors with aborts that will always trip
when things start going south.

This does not add error messages for these cases, this is just a cursory
examination and guesswork.

hpcrun code instead uses hpcrun_terminate, which will call either abort or
monitor_real_abort depending on the environment variable HPCRUN_ABORT_LIBC.
Setting this allows getting stack dumps from the MPI overload on systems
where coredumps are difficult to gather.