Add a new path through the core to mark which Metrics/Scopes have any value
It's technically possible that there will be a Metric with no Statistics,
or where a zero Statistic value does not mean there are no non-zero Metric
values across all Threads. This new data path allows for pruning of unneeded
Contexts from the output, for instance in SparseDB, without making any
assumptions about the Statistics.

Note that fully pruning Contexts from ExperimentXML4 isn't possible yet,
there needs to be a similar flag for timepoint usage as well. It's fairly
trivial to add, so for now this is skipped.