Hotfix to fix the Struct Makefile
In de091ee43ed8deeeda113ee6cbfb5c2730877a52 I changed stringify.sh to support backslashes
in addition to quotes, for FORMATS.md. This broke the Struct Makefile which contains a
handful of double-backslashes (which escape in C).

This commit makes the double backslashes single, so that they get escaped by stringify.sh
and the whole works again.