Elide almost all instruction-type contexts from the meta.db output
Similar to some of the HACKs used for experiment.xml, but far cleaner and
significantly more aggressive in elisions.

The Viewer does not yet support instruction-grain data, eliding it from the output
(hopefully) eases the Viewer transition to the new format. This also means the
output database is unsuitable for anything finer than line-level analysis.

Currently this is hardcoded, once the importance of instruction-grain performance
analysis is decided this can be reverted or adjusted.