Also return the relation's Context when creating Contexts
For example, if a Context expansion generates a Context chain such as
    (point){A}
      (nominal call to):(func){funcB@/b.c:10}  <--
        (enclosed sub-Scope):(line){b.c:12}
          (enclosed sub-Scope):(point){B}      <--
both of the marked Contexts will be returned (previously only the latter
was returned).

This first Context is the relation's Context, it is the Context that
retains the initially requested `(call to)` Relation in the expanded form.
Currently this is always the first created child Context (and the MPI
code currently relies on that fact).

This Context can be used to refer to the relation-edge, rather than the
full nested Context: for instance, function call counts are a property
of the call-edge, this allows for appropriate attribution to the edge
itself (or as close as we can get without changing the data model).