Move applicable metrics from the first instruction to the enclosing function
...Or technically, to the call-edge where the metric was applied. This logic
naturally handles the edge cases (eg. GPUOP on a <gpu copy> placeholder), see the
previous commit for details on what this exactly means.

With this commit, metrics that can only be measured at the kernel level (eg. GKER)
are placed on the kernel function, while proper sampled metrics (eg. REALTIME, GINS)
are placed in the full calling context. No more GKER on some line 0.

This commit (courtesy of @jmellorcrummey) adds a new flag that marks metrics that
need this special handling. This is a stopgap until metrics get a more complete
review, which will be after or during the METRICS.yaml refactor.

As a side-effect, this commit also adjusts the kernel entry Scope (for FlowGraphs)
to be the kernel Function instead of the kernel entry point. It should work, but
PC and range-based sampling support have not been sufficiently tested after making
these refactors. Caveat emptor.

Co-authored-by: John Mellor-Crummey <johnmc@rice.edu>