Fix blocktime (#560)
* Refine hpcviewer manual for 2022.03 release

* Add draft 2022.03 release notes

* Rename to release 2022.04.05

Not sure if it's 2022.03 oe 2022.04. It's likely to be 04.

* Another refinement for man page and manual

* Update hpctoolkit www

* Fix issue #556 (combining gpu and blocktime event)

The root cause of the problem is that we don't allocate properly the event
custom field. It's surprising that it worked for years.

* Refine fix issue #556: remove reference to perf info

Since we use blocktime metric kind, we don't have access to perf info anymore inside the kernel_blocking file.
Remove all access to perf info since we use it for debugging purpose.

* miscellaneous compiler warning cleanup

- remove use of deprecated 'register' keyword
- initialize 'end' time in ppoll (unnecessary)
- add a character to PATH_MAX strings and a terminating NULL

* adjust ompt device support to not fail with Cray cc

* Fix a typo in Args.cpp (#561)

* Fix release notes to comply with the style standard

Minor update to break long lines to conform with style
Minor fix: it's release 2022.04.15, not 2022.04.05

* final updates to documentation

* Move metric_aux_info from cptd to metric descriptor

This additional info contains:
- num_samples: number of samples
- is_multiplexed: flag if the event is multiplexed.
- mean threshold

The 3 fields are used in perf_events to debug or validate the metric values.
At the moment they are calculted manually. See the perf_events implementation.

Reason for code change: to properly fix issue #556
Pros: It's more intuitive, simplifies the access, cleaner code
Cons: Other metric kinds also store this info whether they use it or not.

* Refine blocktime event: convert from nanosec to sec

According to the Linux kernel source code, the unit time is nanosec
See the code of sched_clock()

Although the man-page of perf_events didn't say anything about time granularity,
but I believe it's safe to assume it's always in nanosec.

* Update build instructions and reference packages.yaml.
Remove spack/package.py, it's 2 years old and nothing uses it.

* And actually remove spack/package.py.

* fix EEMSG so that writes to the log occur

* error message if OMPT GPU support has null lookup

* fix issues with renaming structure files in the cache (#563)

- replace "@echo" with "echo" in renamestruct.sh script
  (@echo is a make construct but not legal bash)

- fix monitor_execve failure warning when using hpcrun to measure
  hpcstruct. the issue was that hpcrun was trying to exec
  a script directly. instead of running the rename script
  with command  "<path>/renamestruct.sh", run it with command
  "/bin/sh <path>/renamestruct.sh". that way, /bin/sh can be
  exec'ed. i think that if such an exec failed in the shell,
  the shell probably figures out how to launch the command
  by finding the "#!/bin/sh" in the first line of the shell
  script.

* Fix trace (#564)

* make gpu_trace_fini idempotent to avoid confusion
- both vendor and openmp finalizers call gpu_trace_fini

* ignore threads setting up tracing for an openmp device

* use MONITOR_IGNORE_THREAD with module ignore map

* Fix issue #556 (combining gpu and blocktime event)

The root cause of the problem is that we don't allocate properly the event
custom field. It's surprising that it worked for years.

* Refine fix issue #556: remove reference to perf info

Since we use blocktime metric kind, we don't have access to perf info anymore inside the kernel_blocking file.
Remove all access to perf info since we use it for debugging purpose.

* Move metric_aux_info from cptd to metric descriptor

This additional info contains:
- num_samples: number of samples
- is_multiplexed: flag if the event is multiplexed.
- mean threshold

The 3 fields are used in perf_events to debug or validate the metric values.
At the moment they are calculted manually. See the perf_events implementation.

Reason for code change: to properly fix issue #556
Pros: It's more intuitive, simplifies the access, cleaner code
Cons: Other metric kinds also store this info whether they use it or not.

* Refine blocktime event: convert from nanosec to sec

According to the Linux kernel source code, the unit time is nanosec
See the code of sched_clock()

Although the man-page of perf_events didn't say anything about time granularity,
but I believe it's safe to assume it's always in nanosec.

* change BLOCKTIME units to "sec" from "s"

NOTE: fixing the blocktime and context switch metrics involved reorganizing the implementation of metrics managed by Linux perf_events. After the reorganization, we observe that sometimes the blocktime and context switch metrics are correct. In some cases, when the exclusive time maps to samples within the kernel (with kernel call paths appended below user call paths), the exclusive metrics can be lost. we have decided to accept the PR now and defer full resolution of this issue to prof2.

Co-authored-by: John Mellor-Crummey <johnmc@rice.edu>
Co-authored-by: Keren Zhou <robinho364@gmail.com>
Co-authored-by: Mark W. Krentel <krentel@rice.edu>